refactor:Blacklist와 Customer를 분리하라
